update3

+++ admin/artikel.php
@@ -2,64 +2,68 @@
 include('../koneksi/koneksi.php');
 if((isset($_GET['aksi']))&&(isset($_GET['data']))){
   if($_GET['aksi']=='hapus'){
-     $id_blog = $_GET['data'];
-     $sql_f = "SELECT `cover` FROM `blog` WHERE `id_blog`='$id_blog'";
+     $article_id = $_GET['data'];
+     $sql_f = "SELECT `cover` FROM `articles` WHERE `article_id`='$article_id'";
      $query_f = mysqli_query($koneksi,$sql_f);
-     $jumlah_f = mysqli_num_rows($query_f);
-     if($jumlah_f>0){
-       while($data_f = mysqli_fetch_row($query_f)){
-         $cover = $data_f[0];
-         unlink("cover/$cover");
-       }
+     if($query_f) {
+        $jumlah_f = mysqli_num_rows($query_f);
+        if($jumlah_f > 0){
+          while($data_f = mysqli_fetch_row($query_f)){
+            $cover = $data_f[0];
+            if (file_exists("cover/$cover")) {
+              unlink("cover/$cover");
+            }
+          }
+        }
      }
-   $sql_dh = "delete from `tag_blog` where `id_blog` = '$id_blog'";
+   $sql_dh = "DELETE FROM `tag_articles` WHERE `article_id` = '$article_id'";
    mysqli_query($koneksi,$sql_dh);
-   $sql_dm = "delete from `blog` where `id_blog` = '$id_blog'";
+   $sql_dm = "DELETE FROM `articles` WHERE `article_id` = '$article_id'";
    mysqli_query($koneksi,$sql_dm);
   }
- }
+}
 ?>
 
 <!DOCTYPE html>
- <html>
- <head>
+<html>
+<head>
  <?php include("includes/head.php") ?> 
- </head>
- <body class="hold-transition sidebar-mini layout-fixed">
- <div class="wrapper">
+</head>
+<body class="hold-transition sidebar-mini layout-fixed">
+<div class="wrapper">
  <?php include("includes/header.php") ?>
- 
+
    <?php include("includes/sidebar.php") ?>
- 
+
    <div class="content-wrapper">
      <section class="content-header">
        <div class="container-fluid">
          <div class="row mb-2">
            <div class="col-sm-6">
-             <h3><i class="fab fa-blogger"></i> Blog</h3>
+             <h3><i class="fab fa-blogger"></i> Artikel</h3>
            </div>
            <div class="col-sm-6">
              <ol class="breadcrumb float-sm-right">
-               <li class="breadcrumb-item active"> Blog</li>
+               <li class="breadcrumb-item active"> Artikel</li>
              </ol>
            </div>
          </div>
        </div>
      </section>
- 
+
      <section class="content">
              <div class="card">
                <div class="card-header">
-                 <h3 class="card-title" style="margin-top:5px;"><i class="fas fa-list-ul"></i> Daftar  Blog</h3>
+                 <h3 class="card-title" style="margin-top:5px;"><i class="fas fa-list-ul"></i> Daftar Artikel</h3>
                  <div class="card-tools">
-                   <a href="tambahblog.php" class="btn btn-sm btn-info float-right">
-                   <i class="fas fa-plus"></i> Tambah  Blog</a>
+                   <a href="tambahartikel.php" class="btn btn-sm btn-info float-right">
+                   <i class="fas fa-plus"></i> Tambah Artikel</a>
                  </div>
                </div>
-  
+
                <div class="card-body">
                <div class="col-md-12">
-                   <form method="get" action="blog.php">
+                   <form method="get" action="artikel.php">
                      <div class="row">
                          <div class="col-md-4 bottom-10">
                            <input type="text" class="form-control" id="kata_kunci" name="katakunci">
@@ -104,62 +108,67 @@ if((isset($_GET['aksi']))&&(isset($_GET['data']))){
                           $halaman = $_GET['halaman'];
                           $posisi = ($halaman-1) * $batas;
                      } 
-                      $sql_b = "SELECT `b`.`id_blog`, `b`.`Judul`,     
-                     `k`.`kategori_blog`, DATE_FORMAT(`tanggal`, '%d-%m-%Y') 
-                     FROM `blog` `b`
-                     INNER JOIN `kategori_blog` `k` 
-                     ON `b`.`id_kategori_blog` = `k`.`id_kategori_blog`";
+                     $sql_b = "SELECT `a`.`article_id`, `a`.`title`, `c`.`name`, DATE_FORMAT(`a`.`created_at`, '%d-%m-%Y'), `a`.`cover` 
+                     FROM `articles` `a`
+                     INNER JOIN `categories` `c` 
+                     ON `a`.`category_id` = `c`.`category_id`";
                      if (isset($_GET["katakunci"])) {
                        $katakunci_kategori = $_GET["katakunci"];
-                       $sql_b .= " WHERE `b`.`Judul` LIKE '%$katakunci_kategori%' OR `k`.`kategori_blog` LIKE '%$katakunci_kategori%'";
+                       $sql_b .= " WHERE `a`.`title` LIKE '%$katakunci_kategori%' OR `c`.`name` LIKE '%$katakunci_kategori%'";
                    }                 
-                     $sql_b .= "ORDER BY `k`.`kategori_blog`, `b`.`Judul` limit $posisi, $batas";
+                     $sql_b .= " ORDER BY `c`.`name`, `a`.`title` LIMIT $posisi, $batas";
                      $query_b = mysqli_query($koneksi,$sql_b);
+                     if(!$query_b){
+                         echo "Error: " . mysqli_error($koneksi);
+                         die();
+                     }
                      $no = $posisi+1;
                      while($data_b = mysqli_fetch_row($query_b)){
-                         $id_blog = $data_b[0];
-                         $judul = $data_b[1];
-                         $kategori_blog = $data_b[2];
-                         $tanggal = $data_b[3];
+                         $article_id = $data_b[0];
+                         $title = $data_b[1];
+                         $name = $data_b[2];
+                         $created_at = $data_b[3];
+                         $cover = $data_b[4];
                      ?>
                        <tr>
                          <td><?php echo $no;?></td>
-                         <td><?php echo $kategori_blog;?></td>
-                         <td><?php echo $judul;?></td>
-                         <td><?php echo $tanggal;?></td>
+                         <td><?php echo $name;?></td>
+                         <td><?php echo $title;?></td>
+                         <td><?php echo $created_at;?></td>
                          <td align="center">
-                         <a href="editblog.php?data=<?php echo $id_blog;?>"  
+                         <a href="editartikel.php?data=<?php echo $article_id;?>"  
                          class="btn btn-xs btn-info" title="Edit">
                          <i class="fas fa-edit"></i></a>
-                      <a href="detailblog.php?data=<?php echo $id_blog;?>" 
+                      <a href="viewartikel.php?data=<?php echo $article_id;?>" 
                          class="btn btn-xs btn-info" title="Detail">
                          <i class="fas fa-eye"></i></a>
-                         <a href="javascript:if(confirm('Anda yakin ingin menghapus data <?php echo $judul; ?>?')) 
-                         window.location.href = 'blog.php?aksi=hapus&data=<?php echo $id_blog;?>&notif=hapusberhasil'" 
+                         <a href="javascript:if(confirm('Anda yakin ingin menghapus data <?php echo $title; ?>?')) 
+                         window.location.href = 'artikel.php?aksi=hapus&data=<?php echo $article_id;?>&notif=hapusberhasil'" 
                          class="btn btn-xs btn-warning">
                          <i class="fas fa-trash" title="Hapus">
                          </i></a>                         
                          </td>
                        </tr>
                      <?php $no++;}?>
- 
-                       
                      </tbody>
                    </table>  
                </div>
                
                <?php
-               $sql_jum = "SELECT `b`.`id_blog`, `b`.`Judul`,
-               `k`.`kategori_blog`, DATE_FORMAT(`tanggal`, '%d-%m-%Y')
-               FROM `blog` `b`
-               INNER JOIN `kategori_blog` `k`
-               ON `b`.`id_kategori_blog` = `k`.`id_kategori_blog`";
+               $sql_jum = "SELECT `a`.`article_id`, `a`.`title`, `c`.`name`, DATE_FORMAT(`a`.`created_at`, '%d-%m-%Y')
+               FROM `articles` `a`
+               INNER JOIN `categories` `c`
+               ON `a`.`category_id` = `c`.`category_id`";
                if (isset($_GET["katakunci"])){
                  $katakunci_kategori = $_GET["katakunci"];
-                 $sql_jum .= " where `kategori_blog` LIKE '%$katakunci_kategori%'";
+                 $sql_jum .= " WHERE `c`.`name` LIKE '%$katakunci_kategori%' OR `a`.`title` LIKE '%$katakunci_kategori%'";
                }
-               $sql_jum .= " ORDER BY `k`.`kategori_blog`, `b`.`Judul`";              
+               $sql_jum .= " ORDER BY `c`.`name`, `a`.`title`";              
                $sql_jum = mysqli_query($koneksi,$sql_jum);
+               if(!$sql_jum){
+                   echo "Error: " . mysqli_error($koneksi);
+                   die();
+               }
                $jum_data = mysqli_num_rows($sql_jum);
                $jum_halaman = ceil($jum_data/$batas);
                ?>
@@ -175,24 +184,25 @@ if((isset($_GET['aksi']))&&(isset($_GET['data']))){
                   $setelah = $halaman+1;
                   if($halaman!=1){
                  echo "<li class='page-item'>
-                 <a class='page-link' href='blog.php?halaman=1'>First</a></li>";
+                 <a class='page-link' href='artikel.php?halaman=1'>First</a></li>";
                  echo "<li class='page-item'>
-                 <a class='page-link' href='blog.php?halaman=$sebelum'>«</a></li>";
+                 <a class='page-link' href='artikel.php?halaman=$sebelum'>«</a></li>";
                   }
                 for($i=1; $i<=$jum_halaman; $i++){
-                  if($i!=$halaman){
-                     echo "<li class='page-item'>
-                     <a class='page-link' href='blog.php?halaman=$i'>$i</a></li>";
-                  }else{
-                      echo "<li class='page-item'>
-                      <a class='page-link'>$i</a></li>";
-                  }
+                    if($i>$halaman-5 and $i<$halaman+5){
+                        if($i!=$halaman){
+                            echo "<li class='page-item'><a class='page-link' 
+                            href='artikel.php?halaman=$i'>$i</a></li>";
+                        }else{
+                            echo "<li class='page-item'><a class='page-link'>$i</a></li>";
+                        }
+                    }
                  }
                  if($halaman!=$jum_halaman){
-                      echo "<li class='page-item'>
-                      <a class='page-link' href='blog.php?halaman=$setelah'>»</a></li>";
-                      echo "<li class='page-item'>
-                      <a class='page-link' href='blog.php?halaman=$jum_halaman'>Last</a></li>";
+                 echo "<li class='page-item'>
+                 <a class='page-link' href='artikel.php?halaman=$setelah'>»</a></li>";
+                 echo "<li class='page-item'>
+                 <a class='page-link' href='artikel.php?halaman=$jum_halaman'>Last</a></li>";
                  }
                }
                ?>
